requirements.txt
streamlit
pandas
pdfplumber
tabula-py
openpyxl